Added support to pool of connections
