+++ lib/connection/executor/supervisor.ex
@@ -4,17 +4,8 @@ defmodule Ravix.Connection.RequestExecutor.Supervisor do
 
   require Logger
 
-  alias Ravix.Connection.{RequestExecutor, ServerNode, Topology}
+  alias Ravix.Connection.{RequestExecutor, ServerNode, Topology, NodeSelector}
 
-  @spec init(any) ::
-          {:ok,
-           %{
-             extra_arguments: list,
-             intensity: non_neg_integer,
-             max_children: :infinity | non_neg_integer,
-             period: pos_integer,
-             strategy: :one_for_one
-           }}
   def init(init_arg) do
     DynamicSupervisor.init(
       strategy: :one_for_one,
@@ -34,18 +25,23 @@ defmodule Ravix.Connection.RequestExecutor.Supervisor do
   - store: the store module. E.g: Ravix.Test.Store
   - node: the node to be registered
   """
-  @spec register_node_executor(any, ServerNode.t()) ::
+  @spec register_node_pool(any, ServerNode.t()) ::
           :ignore | {:error, any} | {:ok, pid} | {:ok, pid, any}
-  def register_node_executor(store, %ServerNode{} = node) do
-    Logger.debug("[RAVIX] Registering cluster node '#{node}' for the store '#{inspect(store)}'")
+  def register_node_pool(store, %ServerNode{} = node) do
+    Logger.debug(
+      "[RAVIX] Registering the connection pool with the node '#{node.url}' for the store '#{inspect(store)}'"
+    )
 
     node = %ServerNode{node | store: store}
     DynamicSupervisor.start_child(supervisor_name(store), {RequestExecutor, node})
   end
 
-  @spec remove_node_executor(atom(), pid) :: :ok | {:error, :not_found}
-  def remove_node_executor(store, pid) do
-    Logger.debug("[RAVIX] Removing cluster node '#{inspect(pid)}' for the store '#{inspect(store)}'")
+  @spec remove_node_pool(atom(), pid) :: :ok | {:error, :not_found}
+  def remove_node_pool(store, pid) do
+    Logger.debug(
+      "[RAVIX] Removing cluster node '#{inspect(pid)}' for the store '#{inspect(store)}'"
+    )
+
     DynamicSupervisor.terminate_child(supervisor_name(store), pid)
   end
 
@@ -56,15 +52,21 @@ defmodule Ravix.Connection.RequestExecutor.Supervisor do
   - store: the store module: E.g: Ravix.Test.Store
 
   ## Returns
-  - list({pod, Ravix.Connection.ServerNode})
+  - list(pids)
   """
-  @spec fetch_nodes(any) :: list({pid, ServerNode.t()})
+  def fetch_node_pools(store) do
+    Registry.select(:request_executor_pools, [
+      {{:"$1", :"$2", :"$3"}, [], [{{:"$1", :"$2", :"$3"}}]}
+    ])
+    |> Enum.filter(fn {_pool_name, _pid, node} -> node.store == store end)
+    |> Enum.map(fn {pool_name, pid, node} -> {pid, pool_name, node} end)
+  end
+
+  @spec fetch_nodes(atom()) :: list(pid())
   def fetch_nodes(store) do
-    DynamicSupervisor.which_children(supervisor_name(store))
-    |> Enum.map(fn {_, pid, _kind, _modules} -> pid end)
-    |> Enum.map(fn pid -> {pid, RequestExecutor.fetch_node_state(pid)} end)
-    |> Enum.filter(fn {_, response} -> elem(response, 0) == :ok end)
-    |> Enum.map(fn {pid, {:ok, node}} -> {pid, node} end)
+    Registry.select(:request_executors, [{{:"$1", :"$2", :"$3"}, [], [{{:"$1", :"$2"}}]}])
+    |> Enum.filter(fn {executor_store, _pid} -> executor_store == store end)
+    |> Enum.map(fn {_executor_store, pid} -> pid end)
   end
 
   @doc """
@@ -81,7 +83,7 @@ defmodule Ravix.Connection.RequestExecutor.Supervisor do
           {:new_nodes, list} | {:updated_nodes, list}
         ]
   def update_topology(store, %Topology{} = topology) do
-    current_nodes = fetch_nodes(store)
+    current_nodes = fetch_node_pools(store)
     remaining_nodes = remove_old_nodes(store, current_nodes, topology)
     updated_nodes = update_existing_nodes(remaining_nodes, topology)
     new_nodes = add_new_nodes(store, remaining_nodes, topology)
@@ -90,15 +92,15 @@ defmodule Ravix.Connection.RequestExecutor.Supervisor do
   end
 
   defp remove_old_nodes(store, current_nodes, %Topology{} = topology) do
-    new_nodes_urls = Enum.map(topology.nodes, fn node -> node.url end)
+    new_nodes_ids = Enum.map(topology.nodes, &NodeSelector.node_id(&1))
 
     nodes_to_delete =
       current_nodes
-      |> Enum.reject(fn {_, node} -> Enum.member?(new_nodes_urls, node.url) end)
+      |> Enum.reject(fn {_pid, pool_id, _node} -> Enum.member?(new_nodes_ids, pool_id) end)
 
     nodes_to_delete
-    |> Enum.each(fn {pid, _node} ->
-      RequestExecutor.Supervisor.remove_node_executor(store, pid)
+    |> Enum.each(fn {pid, _, _} ->
+      RequestExecutor.Supervisor.remove_node_pool(store, pid)
     end)
 
     current_nodes -- nodes_to_delete
@@ -106,33 +108,35 @@ defmodule Ravix.Connection.RequestExecutor.Supervisor do
 
   defp update_existing_nodes(nodes, %Topology{} = topology) do
     nodes
-    |> Enum.map(fn {_pid, node} ->
-      [
-        url: node.url,
-        database: node.database,
-        cluster_tag: Topology.cluster_tag_for_node(topology, node.url)
-      ]
+    |> Enum.map(fn {pid, _, node} ->
+      {pid,
+       %ServerNode{
+         node
+         | cluster_tag: Topology.cluster_tag_for_node(topology, node.url)
+       }}
     end)
-    |> Enum.map(fn [url: url, database: database, cluster_tag: cluster_tag] ->
+    |> Enum.map(fn {pid, node} ->
+      RequestExecutor.Supervisor.remove_node_pool(node.store, pid)
+
       Logger.info(
-        "[RAVIX] Updating node '#{inspect(url)}' for the database '#{inspect(database)}' with cluster tag '#{inspect(cluster_tag)}'"
+        "[RAVIX] Updating node '#{inspect(node.url)}' for the database '#{inspect(node.database)}' with cluster tag '#{inspect(node.cluster_tag)}'"
       )
 
-      %{url: url, updated: RequestExecutor.update_cluster_tag(url, database, cluster_tag)}
+      RequestExecutor.Supervisor.register_node_pool(node.store, node)
     end)
   end
 
   defp add_new_nodes(store, existing_nodes, %Topology{} = topology) do
-    existing_nodes_urls = Enum.map(existing_nodes, fn {_pid, node} -> node.url end)
+    existing_nodes = Enum.map(existing_nodes, fn {_pid, pool_id, _node} -> pool_id end)
 
     topology.nodes
-    |> Enum.reject(fn node -> Enum.member?(existing_nodes_urls, node.url) end)
+    |> Enum.reject(&Enum.member?(existing_nodes, NodeSelector.node_id(&1)))
     |> Enum.map(fn new_node ->
       Logger.info(
         "[RAVIX] Registering new node '#{inspect(new_node)}' for the store '#{inspect(store)}'"
       )
 
-      RequestExecutor.Supervisor.register_node_executor(store, new_node)
+      RequestExecutor.Supervisor.register_node_pool(store, new_node)
     end)
   end
