+++ mix.exs
@@ -44,28 +44,29 @@ defmodule Ravix.MixProject do
     [
       {:mint, "~> 1.4"},
       {:jason, "~> 1.3"},
-      {:vex, "~> 0.9.0"},
-      {:mappable, "~> 0.2.4"},
+      {:vex, "~> 0.9"},
+      {:mappable, "~> 0.2"},
       {:elixir_uuid, "~> 1.2"},
       {:ok, "~> 2.3"},
-      {:morphix, "~> 0.8.1"},
+      {:morphix, "~> 0.8"},
       {:timex, "~> 3.7"},
       {:tzdata, "~> 1.1"},
-      {:retry, "~> 0.16.0"},
+      {:retry, "~> 0.16"},
       {:inflex, "~> 2.1"},
-      {:castore, "~> 0.1.17"},
+      {:castore, "~> 0.1"},
+      {:poolboy, "~> 1.5"},
       {:gradient, github: "esl/gradient", only: [:dev, :test], runtime: false},
       {:elixir_sense, github: "elixir-lsp/elixir_sense", only: [:dev]},
-      {:ex_doc, "~> 0.28.3", only: :dev, runtime: false},
-      {:dialyxir, "~> 1.1.0", only: [:dev, :test], runtime: false},
+      {:ex_doc, "~> 0.28", only: :dev, runtime: false},
+      {:dialyxir, "~> 1.1", only: [:dev, :test], runtime: false},
       {:bypass, "~> 2.1", only: :test},
       {:ex_machina, "~> 2.7", only: :test},
-      {:faker, "~> 0.17.0", only: :test},
+      {:faker, "~> 0.17", only: :test},
       {:fake_server, "~> 2.1", only: :test},
-      {:assertions, "~> 0.19.0", only: :test},
-      {:excoveralls, "~> 0.14.4", only: :test},
+      {:assertions, "~> 0.19", only: :test},
+      {:excoveralls, "~> 0.14", only: :test},
       {:credo, "~> 1.6", only: [:dev, :test], runtime: false},
-      {:version_tasks, "~> 0.12.0", only: [:dev, :test], runtime: false}
+      {:version_tasks, "~> 0.12", only: [:dev, :test], runtime: false}
     ]
   end
